+++ fivetran/framework/resources/private_link_test.go
@@ -57,6 +57,38 @@ func setupMockClientPrivateLinkResource(t *testing.T) {
 	)
 }
 
+func setupMockClientPrivateLinkResourceWithNullConfigForImport(t *testing.T) {
+	tfmock.MockClient().Reset()
+	privateLinkResponse :=
+		`{
+        "id": "pl_id",
+        "name": "name",
+        "region": "region",
+        "service": "service",
+        "account_id": "account_id",
+        "cloud_provider": "cloud_provider",
+        "state": "state",
+        "state_summary": "state_summary",
+        "created_at": "created_at",
+        "created_by": "created_by",
+        "host": "host",
+        "config": null
+    }`
+
+	tfmock.MockClient().When(http.MethodGet, "/v1/private-links/pl_id").ThenCall(
+		func(req *http.Request) (*http.Response, error) {
+			privateLinkData = tfmock.CreateMapFromJsonString(t, privateLinkResponse)
+			return tfmock.FivetranSuccessResponse(t, req, http.StatusOK, "", privateLinkData), nil
+		},
+	)
+
+	privateLinkDeleteHandler = tfmock.MockClient().When(http.MethodDelete, "/v1/private-links/pl_id").ThenCall(
+		func(req *http.Request) (*http.Response, error) {
+			return tfmock.FivetranSuccessResponse(t, req, 200, "PrivateLink has been deleted", nil), nil
+		},
+	)
+}
+
 func TestResourcePrivateLinkMock(t *testing.T) {
 	step1 := resource.TestStep{
 		Config: `
@@ -99,6 +131,234 @@ func TestResourcePrivateLinkMock(t *testing.T) {
 
 			Steps: []resource.TestStep{
 				step1,
+			},
+		},
+	)
+}
+
+func TestResourcePrivateLinkImportMock(t *testing.T) {
+	
+	step1 := resource.TestStep{
+		Config: `
+            resource "fivetran_private_link" "test_pl" {
+			provider = fivetran-provider
+
+                name = "name"
+                region = "region"
+                service = "service"
+
+        		config_map = {
+        		  connection_service_name = "connection_service_name"
+        		}
+            }`,
+
+		Check: resource.ComposeAggregateTestCheckFunc(
+			func(s *terraform.State) error {
+				tfmock.AssertEqual(t, privateLinkPostHandler.Interactions, 1)
+				return nil
+			},
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "name", "name"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "region", "region"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "service", "service"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "host", "host"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "config_map.connection_service_name", "connection_service_name"),
+		),
+	}
+
+	step2 := resource.TestStep{
+		ResourceName:      "fivetran_private_link.test_pl",
+		ImportState:       true,
+		ImportStateId: 	"pl_id",
+		ImportStateVerify: true,
+	}
+
+	step3 := resource.TestStep{
+		Config: `
+            resource "fivetran_private_link" "test_pl_imported" {
+			provider = fivetran-provider
+
+                name = "name"
+                region = "region"
+                service = "service"
+            }`,
+		ResourceName:      "fivetran_private_link.test_pl_imported",
+		ImportState:       true,
+		ImportStateId: 	"pl_id",
+		ImportStateCheck: tfmock.ComposeImportStateCheck(
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "id", "pl_id"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "name", "name"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "region", "region"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "service", "service"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "config_map.connection_service_name", "connection_service_name"),
+		),
+	}
+
+	step4 := resource.TestStep{
+		Config: `
+            resource "fivetran_private_link" "test_pl_imported" {
+			provider = fivetran-provider
+
+                name = "name"
+                region = "region"
+                service = "service"
+
+        		config_map = {}
+            }`,
+		ResourceName:      "fivetran_private_link.test_pl_imported",
+		ImportState:       true,
+		ImportStateId: 	"pl_id",
+		ImportStatePersist: true,
+		ImportStateCheck: tfmock.ComposeImportStateCheck(
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "id", "pl_id"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "name", "name"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "region", "region"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "service", "service"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "config_map.connection_service_name", "connection_service_name"),
+		),
+	}
+
+	step5 := resource.TestStep{
+		Config: `
+            resource "fivetran_private_link" "test_pl" {
+			provider = fivetran-provider
+
+                name = "name"
+                region = "region"
+                service = "service"
+
+        		config_map = {
+        		  connection_service_name = "connection_service_name"
+        		}
+            }
+
+            resource "fivetran_private_link" "test_pl_imported" {
+			provider = fivetran-provider
+
+                name = "name"
+                region = "region"
+                service = "service"
+
+        		config_map = {}
+            }`,
+		PlanOnly: true,
+		Check: resource.ComposeAggregateTestCheckFunc (
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "name", "name"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "region", "region"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "service", "service"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "host", "host"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl", "config_map.connection_service_name", "connection_service_name"),
+
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl_imported", "name", "name"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl_imported", "region", "region"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl_imported", "service", "service"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl_imported", "host", "host"),
+			resource.TestCheckResourceAttr("fivetran_private_link.test_pl_imported", "config_map.connection_service_name", "connection_service_name"),
+		),
+	}
+
+	resource.Test(
+		t,
+		resource.TestCase{
+			PreCheck: func() {
+				setupMockClientPrivateLinkResource(t)
+			},
+			ProtoV6ProviderFactories: tfmock.ProtoV6ProviderFactories,
+			CheckDestroy: func(s *terraform.State) error {
+				tfmock.AssertEqual(t, privateLinkDeleteHandler.Interactions, 2)
+				return nil
+			},
+
+			Steps: []resource.TestStep{
+				step1,
+				step2,
+				step3,
+				step4,
+				step5,
+			},
+		},
+	)
+}
+
+func TestResourcePrivateLinkImportWhenConfigIsNullMock(t *testing.T) {
+	
+	step1 := resource.TestStep{
+		Config: `
+            resource "fivetran_private_link" "test_pl_imported" {
+			provider = fivetran-provider
+            }`,
+		ResourceName:      "fivetran_private_link.test_pl_imported",
+		ImportState:       true,
+		ImportStateId: 	"pl_id",
+		ImportStateCheck: tfmock.ComposeImportStateCheck(
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "id", "pl_id"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "name", "name"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "region", "region"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "service", "service"),
+			tfmock.CheckNoImportResourceAttr("fivetran_private_link", "pl_id", "config_map"),
+		),
+	}
+
+	step2 := resource.TestStep{
+		Config: `
+            resource "fivetran_private_link" "test_pl_imported" {
+			provider = fivetran-provider
+
+                name = "name"
+                region = "region"
+                service = "service"
+            }`,
+		ResourceName:      "fivetran_private_link.test_pl_imported",
+		ImportState:       true,
+		ImportStateId: 	"pl_id",
+		ImportStateCheck: tfmock.ComposeImportStateCheck(
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "id", "pl_id"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "name", "name"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "region", "region"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "service", "service"),
+			tfmock.CheckNoImportResourceAttr("fivetran_private_link", "pl_id", "config_map"),
+		),
+	}
+
+	step3 := resource.TestStep{
+		Config: `
+            resource "fivetran_private_link" "test_pl_imported" {
+			provider = fivetran-provider
+
+                name = "name"
+                region = "region"
+                service = "service"
+
+        		config_map = {}
+            }`,
+		ResourceName:      "fivetran_private_link.test_pl_imported",
+		ImportState:       true,
+		ImportStateId: 	"pl_id",
+		ImportStatePersist: true,
+		ImportStateCheck: tfmock.ComposeImportStateCheck(
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "id", "pl_id"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "name", "name"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "region", "region"),
+			tfmock.CheckImportResourceAttr("fivetran_private_link", "pl_id", "service", "service"),
+			tfmock.CheckNoImportResourceAttr("fivetran_private_link", "pl_id", "config_map"),
+		),
+	}
+
+	resource.Test(
+		t,
+		resource.TestCase{
+			PreCheck: func() {
+				setupMockClientPrivateLinkResourceWithNullConfigForImport(t)
+			},
+			ProtoV6ProviderFactories: tfmock.ProtoV6ProviderFactories,
+			CheckDestroy: func(s *terraform.State) error {
+				tfmock.AssertEqual(t, privateLinkDeleteHandler.Interactions, 1)
+				return nil
+			},
+
+			Steps: []resource.TestStep{
+				step1,
+				step2,
+				step3,
 			},
 		},
 	)