+++ fivetran/tests/e2e/resource_private_link_e2e_test.go
@@ -10,6 +10,12 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
 
+const (
+	predefinedPrivateLinkId = "<use the ID of an existing private link>"
+	predefinedPrivateLinkName = "<use the name of an existing private link>"
+
+)
+
 var privateLinkResourceConfig = `
             	resource "fivetran_private_link" "test_pl" {
                 	provider = fivetran-provider
@@ -63,4 +69,118 @@ func testFivetranPrivateLinkResourceCreate(t *testing.T, resourceName string) re
 		//todo: check response _  fields if needed
 		return nil
 	}
+}
+
+func testFivetranPrivateLinkExists(privateLinkId string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		
+		_, err := client.NewPrivateLinkDetails().PrivateLinkId(privateLinkId).Do(context.Background())
+		if err != nil {
+			fmt.Println(err)
+			return err
+		}
+		return nil
+	}
+}
+
+func TestResourcePrivateLinkImportingE2E(t *testing.T) {
+	t.Skip("Private links have a strict limit on the number of entities created. This test should only be used for intermediate tests when changes are made directly to Private links.")
+	
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() {},
+		ProtoV6ProviderFactories: ProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: fmt.Sprintf(`
+					resource "fivetran_private_link" "imported_pl" {
+						provider = fivetran-provider
+
+						name = "%v"
+						region = "AWS_US_EAST_1"
+						service = "REDSHIFT_AWS"
+					}`, predefinedPrivateLinkName),
+				ImportState:            true,
+				ResourceName:            "fivetran_private_link.imported_pl",
+				ImportStateId: predefinedPrivateLinkId,
+			    
+				ImportStateCheck: ComposeImportStateCheck(
+					CheckImportResourceAttr("fivetran_private_link", "id", predefinedPrivateLinkId),
+					CheckImportResourceAttr("fivetran_private_link", "name", predefinedPrivateLinkName),
+					CheckImportResourceAttr("fivetran_private_link", "region", "AWS_US_EAST_1"),
+					CheckImportResourceAttr("fivetran_private_link", "service", "REDSHIFT_AWS"),
+				),
+			},
+			{
+				Config: fmt.Sprintf(`
+					resource "fivetran_private_link" "imported_pl" {
+						provider = fivetran-provider
+
+						name = "%v"
+						region = "AWS_US_EAST_1"
+						service = "REDSHIFT_AWS"
+						config_map = {
+						}
+					}`, predefinedPrivateLinkName),
+				ImportState:            true,
+				ImportStatePersist: 	true,
+				ResourceName:            "fivetran_private_link.imported_pl",
+				ImportStateId: predefinedPrivateLinkId,
+			    
+				ImportStateCheck: ComposeImportStateCheck(
+					CheckImportResourceAttr("fivetran_private_link", "id", predefinedPrivateLinkId),
+					CheckImportResourceAttr("fivetran_private_link", "name", predefinedPrivateLinkName),
+					CheckImportResourceAttr("fivetran_private_link", "region", "AWS_US_EAST_1"),
+					CheckImportResourceAttr("fivetran_private_link", "service", "REDSHIFT_AWS"),
+				),
+			},
+			{
+				Config: fmt.Sprintf(`
+					resource "fivetran_private_link" "imported_pl" {
+						provider = fivetran-provider
+
+						name = "%v"
+						region = "AWS_US_EAST_1"
+						service = "REDSHIFT_AWS"
+						config_map = {
+						}
+					}`, predefinedPrivateLinkName),
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testFivetranPrivateLinkExists(predefinedPrivateLinkId),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "id", predefinedPrivateLinkId),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "name", predefinedPrivateLinkName),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "region", "AWS_US_EAST_1"),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "service", "REDSHIFT_AWS"),
+				),
+			},
+			{
+				Config: fmt.Sprintf(`
+					resource "fivetran_private_link" "imported_pl" {
+						provider = fivetran-provider
+
+						name = "%v"
+						region = "AWS_US_EAST_1"
+						service = "REDSHIFT_AWS"
+						config_map = {}
+					}`, predefinedPrivateLinkName),
+				PlanOnly: true,
+			},
+			{
+				RefreshState: true,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testFivetranPrivateLinkExists(predefinedPrivateLinkId),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "id", predefinedPrivateLinkId),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "name", predefinedPrivateLinkName),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "region", "AWS_US_EAST_1"),
+					resource.TestCheckResourceAttr("fivetran_private_link.imported_pl", "service", "REDSHIFT_AWS"),
+				),
+			},
+			{
+				Config: `
+				`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					testFivetranPrivateLinkExists(predefinedPrivateLinkId),
+				),
+			},
+		},
+	})
 }