+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.9.14...HEAD)
+## [Unreleased](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.9.15...HEAD)
+
+## [v1.9.15](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.9.14...v1.9.15)
+
+### Fixed
+- Allow for importing of `fivetran_private_link` with empty `config_map` (`config_map = {}`).
 
 ## [v1.9.14](https://github.com/fivetran/terraform-provider-fivetran/compare/v1.9.13...v1.9.14)